Implemented Java Collections (List, Set, Map) features in EMS

+++ src/main/java/com/employee/repository/LoginRepository.java
@@ -5,7 +5,13 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface LoginRepository extends JpaRepository<LoginDetails,Integer> {
+import java.util.Optional;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.employee.entity.LoginDetails;
+
+public interface LoginRepository extends JpaRepository<LoginDetails, Long> {
 
     Optional<LoginDetails> findByUsername(String username);
 }
